Added file _version.py - holds the current version of the project.
Update the licence ot V3 of the GNU General Public License.

Added files that have changed size and date to file that are copied
across.
Added a reverse scaan, that will delete file in the destination that no
longer exist in the source directory.

Added command argument to print the license and version.

Change command argument for source and destination to be optional, but
checks that they are provided if needed.